+++ release-notes.md
@@ -3,7 +3,7 @@
 This is a major release to support Helm v3.
 It is recommended you read the [Helm 3 migration guide](https://helm.sh/docs/topics/v2_v3_migration/) before using this release.
 
-> Starting from this release, support for Helmsman v1.x will be limited to bug fixes.
+> Starting from Helmsman v3.0.0 GA release, support for Helmsman v1.x will be limited to bug fixes.
 
 The following are the most important changes:
 - A new and improved logger.
@@ -13,6 +13,7 @@ The following are the most important changes:
 - Deprecating all the DSF stanzas related to Tiller.
 - Deprecating the `purge` option for releases.
 - The default value for `storageBackend` is now `secret`.
+- The `stable` and `incubator` repos are no longer added by default and the `--no-default-repos` flag is deprecated.
 - The `--suppress-diff-secrets` cmd flag is enabled by default.
 - The `--no-env-values-subst` cmd flag is enabled by default.
 - The `--no-ssm-values-subst` cmd flag is enabled by default.